kodi - only add the pippl repos for Raspberry Pi OS (buster) and lower

+++ scriptmodules/ports/kodi.sh
@@ -21,16 +21,18 @@ function _update_hook_kodi() {
 }
 
 function depends_kodi() {
-    # raspbian
+    # Raspberry Pi OS
     if [[ "$__os_id" = "Raspbian" ]] && isPlatform "rpi"; then
-        if [[ "$md_mode" == "install" ]]; then
-            # remove old repository
-            rm -f /etc/apt/sources.list.d/mene.list
-            echo "deb http://pipplware.pplware.pt/pipplware/dists/$__os_codename/main/binary/ ./" >/etc/apt/sources.list.d/pipplware.list
-            download http://pipplware.pplware.pt/pipplware/key.asc - | apt-key add - &>/dev/null
-        else
-            rm -f /etc/apt/sources.list.d/pipplware.list
-            apt-key del 4096R/BAA567BB >/dev/null
+        if [[ "$__os_debian_ver" -le 10 ]]; then
+            if [[ "$md_mode" == "install" ]]; then
+                # remove old repository
+                rm -f /etc/apt/sources.list.d/mene.list
+                echo "deb http://pipplware.pplware.pt/pipplware/dists/$__os_codename/main/binary/ ./" >/etc/apt/sources.list.d/pipplware.list
+                download http://pipplware.pplware.pt/pipplware/key.asc - | apt-key add - &>/dev/null
+            else
+                rm -f /etc/apt/sources.list.d/pipplware.list
+                apt-key del 4096R/BAA567BB >/dev/null
+            fi
         fi
     # ubuntu
     elif [[ -n "$__os_ubuntu_ver" ]] && isPlatform "x86"; then
